requirements.txt
django==1.11.8
django-ranged-response==0.2.0
django-redis==4.8.0
django-simple-captcha==0.5.6
Pillow==4.2.1
redis==2.10.6
supervisor==3.3.
PyCrypto